Add address and donate button fields to footer
Introduces 'address' and 'donateButton' fields to the footer config and types, allowing dynamic display of the address and a configurable donate button in the footer. Updates the Footer component to use these new fields and disables the donate button if no URL is provided. Also removes usage of a non-existent 'description' field from the ProjectsSection component.